requirements.txt
django-3.2.4
djangorestframework-3.12.4